Update Layout component to include profile.
