--- src/pages/layout.js
@@ -1,57 +1,0 @@
-import banner from "../../assets/banner.png?as=webp";
-import { Link, Outlet } from "react-router-dom";
-import NavItem from "../components/nav_item";
-import React, { useEffect } from "react";
-import { useLocation } from "react-router-dom";
-
-const Layout = () => {
-  const img = (
-    <img
-      src={banner}
-      className="drop-shadow-md rounded-lg w-full"
-      alt="Hall generated by Midjourney"
-    />
-  );
-
-  return (
-    <main className="container font-roboto-slab leading-6 max-w-3xl mx-auto px-6 pt-12 pb-16 text-white w-full">
-      <header>
-        {isOnPage("/") ? img : <Link to="/">{img}</Link>}
-
-        <nav className="px-4 pt-6 pb-8">
-          <NavItem
-            emoji="👋"
-            isFirst
-            onPage={isOnPage("/")}
-            path="/"
-            text="About"
-          />
-          <NavItem
-            emoji="🏗"
-            onPage={isOnPage("/projects")}
-            path="/projects"
-            text="Projects"
-          />
-          <NavItem emoji="📇" onPage={isOnPage("/cv")} path="/cv" text="CV" />
-          <NavItem
-            emoji="🍕"
-            onPage={isOnPage("/now")}
-            path="/now"
-            text="Now"
-          />
-        </nav>
-      </header>
-
-      <section className="px-4">
-        <Outlet />
-      </section>
-    </main>
-  );
-};
-
-const isOnPage = (path) => {
-  const location = useLocation();
-  return location.pathname === path;
-};
-
-export default Layout;
